fix: changed dto return types and added env

+++ WeatherForecast.Application/Contracts/Dtos/SummaryDto.cs
@@ -9,7 +9,7 @@ namespace WeatherForecast.Application.Contracts.Dtos
     public class SummaryDto
     {
         public float AveragePressure { get; set; }
-        public float AverageSunshineTime { get; set; }
+        public int AverageSunshineTime { get; set; }
         public float TemperatureMin { get; set; }
         public float TemperatureMax { get; set; }
         public string WeekSummary { get; set; } = string.Empty;
